requirements.txt
fastapi==0.115.6
httpx==0.28.1
Jinja2==3.1.4
markdown2==2.5.2
nicegui==2.8.1
pydantic==2.10.3
python-dotenv==1.0.1
PyYAML==6.0.2
uvicorn==0.34.0
numpy==2.2.1
